Add files via upload
As a part of the data science team in the organization, the task is to analyze the 
financial metrics of different companies. and develop a predictive model leveraging 
machine learning techniques to identify whether a given company will default on its 
debt repayments in the next two quarters. The predictive model will help the 
organization anticipate potential challenges with the financial performance of the 
companies and enable proactive risk mitigation strategies.